DerelictLUA fixes.
Fixed a bug that should have been caught before, but was only being flagged by DMD 2.061 beta when compiling to 32-bit on Linux. lua_newState was declared to return a lua_State by value, rather than by pointer. Which should be rather impossible given that it's not defined in D except as an opaque struct.
Fixed some formatting annoyances in lua.functions, where the * was next to the variable name rather than the type.
Removed the opaque struct declaration of va_list from lua.types. It had no business there. Plus, it was the wrong sort of declarationa and, again, shouldn't have compiled since it was used as a parameter in lua.funcitons without a *.
Added an import of core.stdc.stdarg to lua.functions for the proper declaration of va_list.